src/requirements.txt
Django==1.10.4
django-extensions==1.7.7

django-disposable-email-checker==1.2.1

# PostgresSQL

# serve files
whitenoise==3.2.2

# Rest Framework
djangorestframework==3.5.4

# Sentry
raven==5.27.1

# for date processing
python-dateutil==2.6.0

# Logging
git+https://github.com/Seedstars/django-rest-logger.git

django-rest-knox==3.0.0

mock==2.0.0
factory-boy==2.8.1

prospector==0.12.4
bandit==1.4.0

pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.0.0
pytest-pythonpath==0.7.1
pytest-xdist==1.15.0

ptpython==0.36
MySQL-python==1.2.5

celery==4.0.2
demjson==2.2.4
django-celery==3.1.17
django-redis==4.6.0
django-redis-cache==1.7.1
redis==2.10.5
pytz==2016.10
rrdtool==0.1.11
scapy==2.3.3
django-cors-headers==2.0.2
